+++ docs/source/releasenotes.rst
@@ -27,6 +27,12 @@ This release makes a couple of potentially breaking changes:
   renamed to :code:`start_line_pos` and :code:`start_line_no`, to distinguish
   them from the new fields starting :code:`end_*`.
 
+* When linting from stdin, if there are no violations found - before this version,
+  the serialised response would be simply an empty list (:code:`[]`). From 3.0
+  onwards, there will now  be a record for the *file* with some statistics,
+  but the *violations* section of the response for that file will still be an
+  empty list.
+
 * The default :code:`annotation_level` set by the :code:`--annotation-level`
   option on the :code:`sqlfluff lint` command has been changed from :code:`notice`
   to :code:`warning`, to better distinguish linting errors from warnings, which